+++ src/compile/lower.rs
@@ -3,7 +3,7 @@
 //! This module is responsible for taking Expressions and lowering
 //! them to LLVM.
 
-use syntax::{Expression, Constant};
+use syntax::ast::expression::{Expression, Constant};
 use syntax::ast::operators::{PrefixOp, InfixOp};
 use low_loader::prelude::*;
