+++ src/syntax/ast/mod.rs
@@ -1,1 +1,9 @@
 pub mod operators;
+pub mod types;
+pub mod expression;
+
+pub mod prelude {
+    pub use super::expression::Expression;
+    pub use super::types::{TypeRef, TypedId};
+    pub use super::operators::{InfixOp, PrefixOp};
+}
